added support for multiple time formats
new birdflow_options() has options for time_format and verbose
output structures generally  now format the time based on the option
and (new) preserve the time associated with a single distribution
by saving it  to a "time" attribute.  Functions that create SpatRast
objects preserve the time in the layer name.  Also fixed #33,
import_birdflow() now properly labels the columns of the distributions,
so timestep (or other time format) is preserved or generated on output
from get_distr().